Split ci_processing into smaller parts.
Splitting methods into modules. Each module includes code specific for a
single button action.
