+++ app/controllers/application_controller/ci_processing.rb
@@ -2,971 +2,16 @@ module ApplicationController::CiProcessing
   extend ActiveSupport::Concern
 
   included do
-    private(:process_elements)
+    include Mixins::Actions::VmActions::Ownership
+    include Mixins::Actions::VmActions::Retire
+    include Mixins::Actions::VmActions::LiveMigrate
+    include Mixins::Actions::VmActions::Evacuate
+    include Mixins::Actions::VmActions::AssociateFloatingIp
+    include Mixins::Actions::VmActions::DisassociateFloatingIp
+    include Mixins::Actions::VmActions::Resize
+    include Mixins::Actions::VmActions::RightSize
+    include Mixins::Actions::VmActions::Reconfigure
     helper_method :supports_reconfigure_disks?
-  end
-
-  def ownership_form_fields
-    render :json => build_ownership_hash(params[:object_ids])
-  end
-
-  # Set Ownership selected db records
-  def set_ownership
-    assert_privileges(params[:pressed])
-    ownership_items = []
-    # check to see if coming from show_list or drilled into vms from another CI
-    controller = if request.parameters[:controller] == "vm" || ["all_vms", "vms", "instances", "images"].include?(params[:display])
-                   "vm"
-                 elsif ["miq_templates", "images"].include?(params[:display]) || params[:pressed].starts_with?("miq_template_")
-                   "miq_template"
-                 else
-                   request.parameters[:controller]
-                 end
-    @edit ||= {}
-    @edit[:controller] = controller
-    recs = []
-    if !session[:checked_items].nil? && @lastaction == "set_checked_items"
-      recs = session[:checked_items]
-    else
-      recs = find_checked_ids_with_rbac(get_class_from_controller_param(controller))
-    end
-    if recs.blank?
-      id = find_id_with_rbac(get_class_from_controller_param(params[:controller]), params[:id])
-      recs = [id.to_i]
-    end
-    @edit[:object_ids] = recs
-    if recs.length < 1
-      add_flash(_("One or more %{model} must be selected to Set Ownership") % {
-        :model => Dictionary.gettext(db.to_s, :type => :model, :notfound => :titleize, :plural => true)}, :error)
-      @refresh_div = "flash_msg_div"
-      @refresh_partial = "layouts/flash_msg"
-      return
-    else
-      ownership_items = recs.collect(&:to_i)
-    end
-
-    if filter_ownership_items(get_class_from_controller_param(controller), ownership_items).empty?
-      add_flash(_('None of the selected items allow ownership changes'), :error)
-      @refresh_div = "flash_msg_div"
-      @refresh_partial = "layouts/flash_msg"
-      return
-    end
-
-    if @explorer
-      @sb[:explorer] = true
-      ownership(ownership_items)
-    else
-      if role_allows?(:feature => "vm_ownership")
-        drop_breadcrumb(:name => _("Set Ownership"), :url => "/vm_common/ownership")
-        javascript_redirect :controller => controller, :action => 'ownership', :rec_ids => ownership_items, :escape => false # redirect to build the ownership screen
-      else
-        head :ok
-      end
-    end
-  end
-
-  alias_method :image_ownership, :set_ownership
-  alias_method :instance_ownership, :set_ownership
-  alias_method :vm_ownership, :set_ownership
-  alias_method :miq_template_ownership, :set_ownership
-
-  def get_class_from_controller_param(controller)
-    case controller
-    when "orchestration_stack"
-      OrchestrationStack
-    when "service"
-      Service
-    when "vm_or_template", "vm_infra", "vm_cloud", "vm"
-      VmOrTemplate
-    when "miq_template"
-      MiqTemplate
-    end
-  end
-
-  # Assign/unassign ownership to a set of objects
-  def ownership(ownership_items = [])
-    @sb[:explorer] = true if @explorer
-    @in_a_form = @ownershipedit = true
-    drop_breadcrumb(:name => _("Set Ownership"), :url => "/vm_common/ownership")
-    ownership_items = params[:rec_ids] if params[:rec_ids]
-    build_ownership_info(ownership_items)
-    return if @ownershipitems.empty?
-    build_targets_hash(@ownershipitems)
-    if @sb[:explorer]
-      @refresh_partial = "shared/views/ownership"
-    else
-      render :action => "show"
-    end
-  end
-
-  DONT_CHANGE_OWNER = "0"
-
-  def filter_ownership_items(klass, ownership_items)
-    @origin_ownership_items = ownership_items
-    @ownershipitems ||= begin
-      ownership_scope = klass.where(:id => ownership_items)
-      ownership_scope = ownership_scope.with_ownership if klass.respond_to?(:with_ownership)
-      Rbac.filtered(ownership_scope.order(:name))
-    end
-  end
-
-  def build_ownership_info(ownership_items)
-    @edit ||= {}
-    klass = get_class_from_controller_param(params[:controller])
-    record = klass.find(ownership_items[0])
-    user = record.evm_owner if ownership_items.length == 1
-    @user = user ? user.id.to_s : nil
-
-    @groups = {} # Create new entries hash (2nd pulldown)
-    # need to do this only if 1 vm is selected and miq_group has been set for it
-    group = record.miq_group if ownership_items.length == 1
-    @group = group ? group.id.to_s : nil
-    Rbac.filtered(MiqGroup.non_tenant_groups).each { |g| @groups[g.description] = g.id.to_s }
-
-    @user = @group = DONT_CHANGE_OWNER if ownership_items.length > 1
-    @edit[:object_ids] = filter_ownership_items(klass, ownership_items)
-    @view = get_db_view(klass == VmOrTemplate ? Vm : klass) # Instantiate the MIQ Report view object
-    @view.table = MiqFilter.records2table(@ownershipitems, @view.cols + ['id'])
-    session[:edit] = @edit
-  end
-
-  # Build the ownership assignment screen
-  def build_ownership_hash(ownership_items)
-    klass = get_class_from_controller_param(params[:controller])
-    record = klass.find(ownership_items[0])
-    user = record.evm_owner if ownership_items.length == 1
-    @user = user ? user.id.to_s : ''
-    @groups = {}
-    group = record.miq_group if ownership_items.length == 1
-    @group = group ? group.id.to_s : nil
-    Rbac.filtered(MiqGroup).each { |g| @groups[g.description] = g.id.to_s }
-    @user = @group = DONT_CHANGE_OWNER if ownership_items.length > 1
-    @ownershipitems = Rbac.filtered(klass.where(:id => ownership_items).order(:name), :class => klass)
-    raise _('Invalid items passed') unless @ownershipitems.pluck(:id).to_set == ownership_items.map(&:to_i).to_set
-    {:user  => @user,
-     :group => @group}
-  end
-
-  def valid_items_for(klass, param_ids)
-    scope = klass.respond_to?(:with_ownership) ? klass.with_ownership : klass
-    checked_ids = Rbac.filtered(scope.where(:id => param_ids)).pluck(:id)
-    checked_ids.to_set == param_ids.to_set
-  end
-
-  def ownership_update
-    case params[:button]
-    when "cancel"
-      add_flash(_("Set Ownership was cancelled by the user"))
-      if @sb[:explorer]
-        @edit = @sb[:action] = nil
-        replace_right_cell
-      else
-        session[:flash_msgs] = @flash_array
-        javascript_redirect previous_breadcrumb_url
-      end
-    when "save"
-      opts = {}
-      unless params[:user] == DONT_CHANGE_OWNER
-        if params[:user].blank?     # to clear previously set user
-          opts[:owner] = nil
-        elsif params[:user] != @user
-          opts[:owner] = User.find(params[:user])
-        end
-      end
-
-      unless params[:group] == DONT_CHANGE_OWNER
-        if params[:group].blank?    # to clear previously set group
-          opts[:group] = nil
-        elsif params[:group] != @group
-          opts[:group] = MiqGroup.find_by_id(params[:group])
-        end
-      end
-
-      klass = get_class_from_controller_param(request.parameters[:controller])
-      param_ids = params[:objectIds].map(&:to_i)
-      raise _('Invalid items selected.') unless valid_items_for(klass, param_ids)
-
-      result = klass.set_ownership(param_ids, opts)
-      unless result == true
-        result["missing_ids"].each { |msg| add_flash(msg, :error) } if result["missing_ids"]
-        result["error_updating"].each { |msg| add_flash(msg, :error) } if result["error_updating"]
-        javascript_flash
-      else
-        object_types = object_types_for_flash_message(klass, params[:objectIds])
-
-        flash = _("Ownership saved for selected %{object_types}") % {:object_types => object_types}
-        add_flash(flash)
-        if @sb[:explorer]
-          @sb[:action] = nil
-          replace_right_cell
-        else
-          session[:flash_msgs] = @flash_array
-          javascript_redirect previous_breadcrumb_url
-        end
-      end
-    when "reset"
-      @in_a_form = true
-      if @edit[:explorer]
-        ownership
-        add_flash(_("All changes have been reset"), :warning)
-        request.parameters[:controller] == "service" ? replace_right_cell(:nodetype => "ownership") : replace_right_cell
-      else
-        javascript_redirect :action        => 'ownership',
-                            :flash_msg     => _("All changes have been reset"),
-                            :flash_warning => true,
-                            :escape        => true
-      end
-    end
-  end
-
-  # Retire 1 or more items (vms, stacks, services)
-  def retirevms
-    assert_privileges(params[:pressed])
-    # check to see if coming from show_list or drilled into vms from another CI
-    if request.parameters[:controller] == "vm" || %w(all_vms instances vms).include?(params[:display])
-      rec_cls = "vm"
-      bc_msg = _("Retire VM or Instance")
-    elsif request.parameters[:controller] == "service"
-      rec_cls =  "service"
-      bc_msg = _("Retire Service")
-    elsif request.parameters[:controller] == "orchestration_stack" || %w(orchestration_stacks).include?(params[:display])
-      rec_cls = "orchestration_stack"
-      bc_msg = _("Retire Orchestration Stack")
-    end
-    klass = rec_cls ? rec_cls.camelize.constantize : get_class_from_controller_param(params[:controller])
-    selected_items = find_checked_ids_with_rbac(klass)
-    @edit ||= {}
-    @edit[:object_ids] = selected_items
-    session[:edit] = @edit
-    if !%w(orchestration_stack service).include?(request.parameters["controller"]) && !%w(orchestration_stacks).include?(params[:display]) &&
-       VmOrTemplate.find(selected_items).any? { |vm| !vm.supports_retire? }
-      add_flash(_("Set Retirement Date does not apply to selected %{model}") %
-        {:model => ui_lookup(:table => "miq_template")}, :error)
-      javascript_flash(:scroll_top => true)
-      return
-    end
-    if selected_items.blank?
-      session[:retire_items] = [params[:id]]
-    elsif selected_items.empty?
-        add_flash(_("At least one %{model} must be selected for tagging") %
-          {:model => ui_lookup(:model => "Vm")}, :error)
-        @refresh_div = "flash_msg_div"
-        @refresh_partial = "layouts/flash_msg"
-        return
-    else
-      session[:retire_items] = selected_items # Set the array of retire items
-    end
-    session[:assigned_filters] = assigned_filters
-    if @explorer
-      retire
-    else
-      drop_breadcrumb(:name => bc_msg,
-                      :url  => "/#{session[:controller]}/retire")
-      javascript_redirect :controller => rec_cls, :action => 'retire' # redirect to build the retire screen
-    end
-  end
-  alias_method :instance_retire, :retirevms
-  alias_method :vm_retire, :retirevms
-  alias_method :orchestration_stack_retire, :retirevms
-
-  def retirement_info
-    case request.parameters[:controller]
-    when "orchestration_stack"
-      assert_privileges("orchestration_stack_retire")
-      kls = OrchestrationStack
-    when "service"
-      assert_privileges("service_retire")
-      kls = Service
-    when "vm_cloud", "vm"
-      assert_privileges("instance_retire")
-      kls = Vm
-    when "vm_infra"
-      assert_privileges("vm_retire")
-      kls = Vm
-    end
-    obj = kls.find_by_id(params[:id])
-    render :json => {
-      :retirement_date    => obj.retires_on.try(:strftime, '%m/%d/%Y'),
-      :retirement_warning => obj.retirement_warn
-    }
-  end
-
-  # Build the retire VMs screen
-  def retire
-    @sb[:explorer] = true if @explorer
-    kls = case request.parameters[:controller]
-          when "orchestration_stack"
-            OrchestrationStack
-          when "service"
-            Service
-          when "vm_infra", "vm_cloud", "vm", "vm_or_template"
-            Vm
-          end
-    if params[:button]
-      if params[:button] == "cancel"
-        flash = _("Set/remove retirement date was cancelled by the user")
-        @sb[:action] = nil
-      elsif params[:button] == "save"
-        if params[:retire_date].blank?
-          t = nil
-          w = nil
-
-          if session[:retire_items].length == 1
-            flash = _("Retirement date removed")
-          else
-            flash = _("Retirement dates removed")
-          end
-        else
-          t = params[:retire_date].in_time_zone
-          w = params[:retire_warn].to_i
-
-          ts = t.strftime("%x %R %Z")
-          if session[:retire_items].length == 1
-            flash = _("Retirement date set to %{date}") % {:date => ts}
-          else
-            flash = _("Retirement dates set to %{date}") % {:date => ts}
-          end
-        end
-        kls.retire(session[:retire_items], :date => t, :warn => w) # Call the model to retire the VM(s)
-        @sb[:action] = nil
-      end
-      add_flash(flash)
-      if @sb[:explorer]
-        replace_right_cell
-      else
-        session[:flash_msgs] = @flash_array.dup
-        javascript_redirect previous_breadcrumb_url
-      end
-      return
-    end
-    session[:changed] = @changed = false
-    drop_breadcrumb(:name => _("Retire %{name}") % {:name => ui_lookup(:models => kls.to_s)},
-                    :url  => "/#{session[:controller]}/retire")
-    session[:cat] = nil                 # Clear current category
-    @retireitems = kls.find(session[:retire_items]).sort_by(&:name) # Get the db records
-    build_targets_hash(@retireitems)
-    @view = get_db_view(kls)              # Instantiate the MIQ Report view object
-    @view.table = MiqFilter.records2table(@retireitems, @view.cols + ['id'])
-    if @retireitems.length == 1 && !@retireitems[0].retires_on.nil?
-      t = @retireitems[0].retires_on                                         # Single VM, set to current time
-      w = @retireitems[0].retirement_warn if @retireitems[0].retirement_warn # Single VM, get retirement warn
-    else
-      t = nil
-    end
-    session[:retire_date] = t.nil? ? nil : "#{t.month}/#{t.day}/#{t.year}"
-    session[:retire_warn] = w
-    @in_a_form = true
-    @edit ||= {}
-    @edit[:object_ids] = @retireitems
-    session[:edit] = @edit
-    @refresh_partial = "shared/views/retire" if @explorer || @layout == "orchestration_stack"
-  end
-
-  def resizevms
-    assert_privileges("instance_resize")
-    recs = find_checked_items
-    recs = [params[:id].to_i] if recs.blank?
-    @record = find_record_with_rbac(VmOrTemplate, recs.first) # Set the VM object
-    if @record.supports_resize?
-      if @explorer
-        resize
-        @refresh_partial = "vm_common/resize"
-      else
-        javascript_redirect :controller => 'vm', :action => 'resize', :rec_id => @record.id, :escape => false # redirect to build the retire screen
-      end
-    else
-      add_flash(_("Unable to reconfigure %{instance} \"%{name}\": %{details}") % {
-        :instance => ui_lookup(:table => 'vm_cloud'),
-        :name     => @record.name,
-        :details  => @record.unsupported_reason(:resize)}, :error)
-    end
-  end
-  alias instance_resize resizevms
-
-  def resize
-    assert_privileges("instance_resize")
-    @record ||= find_record_with_rbac(VmOrTemplate, params[:rec_id])
-    unless @explorer
-      drop_breadcrumb(
-        :name => _("Reconfigure Instance '%{name}'") % {:name => @record.name},
-        :url  => "/vm/resize"
-      )
-    end
-    @sb[:explorer] = @explorer
-    @in_a_form = true
-    @resize = true
-    render :action => "show" unless @explorer
-  end
-
-  def resize_vm
-    assert_privileges("instance_resize")
-    @record = find_record_with_rbac(VmOrTemplate, params[:id])
-
-    case params[:button]
-    when "cancel"
-      add_flash(_("Reconfigure of %{model} \"%{name}\" was cancelled by the user") % {
-        :model => ui_lookup(:table => "vm_cloud"), :name => @record.name})
-      @record = @sb[:action] = nil
-    when "submit"
-      if @record.supports_resize?
-        begin
-          flavor_id = params['flavor_id']
-          flavor = find_record_with_rbac(Flavor, flavor_id)
-          old_flavor_name = @record.flavor.try(:name) || _("unknown")
-          @record.resize_queue(session[:userid], flavor)
-          add_flash(_("Reconfiguring %{instance} \"%{name}\" from %{old_flavor} to %{new_flavor}") % {
-            :instance   => ui_lookup(:table => 'vm_cloud'),
-            :name       => @record.name,
-            :old_flavor => old_flavor_name,
-            :new_flavor => flavor.name})
-        rescue => ex
-          add_flash(_("Unable to reconfigure %{instance} \"%{name}\": %{details}") % {
-            :instance => ui_lookup(:table => 'vm_cloud'),
-            :name     => @record.name,
-            :details  => get_error_message_from_fog(ex.to_s)}, :error)
-        end
-      else
-        add_flash(_("Unable to reconfigure %{instance} \"%{name}\": %{details}") % {
-          :instance => ui_lookup(:table => 'vm_cloud'),
-          :name     => @record.name,
-          :details  => @record.unsupported_reason(:resize)}, :error)
-      end
-      params[:id] = @record.id.to_s # reset id in params for show
-      @record = nil
-      @sb[:action] = nil
-    end
-    if @sb[:explorer]
-      replace_right_cell
-    else
-      session[:flash_msgs] = @flash_array.dup
-      javascript_redirect previous_breadcrumb_url
-    end
-    return
-  end
-
-  def resize_form_fields
-    assert_privileges("instance_resize")
-    @record = find_record_with_rbac(VmOrTemplate, params[:id])
-    flavors = []
-    unless @record.ext_management_system.nil?
-      @record.ext_management_system.flavors.each do |ems_flavor|
-        # include only flavors with root disks at least as big as the instance's current root disk.
-        if @record.flavor.nil? || ((ems_flavor != @record.flavor) && (ems_flavor.root_disk_size >= @record.flavor.root_disk_size))
-          flavors << {:name => ems_flavor.name_with_details, :id => ems_flavor.id}
-        end
-      end
-    end
-    render :json => {
-      :flavors => flavors
-    }
-  end
-
-  def livemigratevms
-    assert_privileges("instance_live_migrate")
-    recs = find_checked_ids_with_rbac(VmOrTemplate)
-    recs = [params[:id].to_i] if recs.blank?
-    session[:live_migrate_items] = recs
-    if @explorer
-      live_migrate
-      @refresh_partial = "vm_common/live_migrate"
-    else
-      javascript_redirect :controller => 'vm', :action => 'live_migrate', :escape => false
-    end
-  end
-  alias instance_live_migrate livemigratevms
-
-  def live_migrate
-    assert_privileges("instance_live_migrate")
-    drop_breadcrumb(
-      :name => _("Live Migrate Instances"),
-      :url  => "/vm_cloud/live_migrate"
-    ) unless @explorer
-    @sb[:explorer] = @explorer
-    @in_a_form = true
-    @live_migrate = true
-
-    @live_migrate_items = VmOrTemplate.find(session[:live_migrate_items]).sort_by(&:name)
-    build_targets_hash(@live_migrate_items)
-    @view = get_db_view(VmOrTemplate)
-    @view.table = MiqFilter.records2table(@live_migrate_items, @view.cols + ['id'])
-
-    render :action => "show" unless @explorer
-  end
-
-  def live_migrate_form_fields
-    assert_privileges("instance_live_migrate")
-    @record = find_record_with_rbac(VmOrTemplate, params[:id])
-    hosts = []
-    unless @record.ext_management_system.nil?
-      # wrap in a rescue block in the event the connection to the provider fails
-      begin
-        connection = @record.ext_management_system.connect
-        current_hostname = connection.handled_list(:servers).find do |s|
-          s.name == @record.name
-        end.os_ext_srv_attr_hypervisor_hostname
-        # OS requires its own name for the host be used in the migrate API, so get the
-        # provider hostname from fog.
-        hosts = connection.hosts.select { |h| h.service_name == "compute" && h.host_name != current_hostname }.map do |h|
-          {:name => h.host_name, :id => h.host_name}
-        end
-      rescue
-        hosts = []
-      end
-    end
-    render :json => {
-      :hosts => hosts
-    }
-  end
-
-  def live_migrate_vm
-    assert_privileges("instance_live_migrate")
-    case params[:button]
-    when "cancel"
-      add_flash(_("Live migration of Instances was cancelled by the user"))
-    when "submit"
-      @live_migrate_items = VmOrTemplate.find(session[:live_migrate_items]).sort_by(&:name)
-      @live_migrate_items.each do |vm|
-        if vm.supports_live_migrate?
-          options = {
-            :hostname         => params['auto_select_host'] == 'on' ? nil : params['destination_host'],
-            :block_migration  => params['block_migration']   == 'on',
-            :disk_over_commit => params['disk_over_commit']  == 'on'
-          }
-          task_id = vm.class.live_migrate_queue(session[:userid], vm, options)
-          unless task_id.kind_of?(Integer)
-            add_flash(_("Instance live migration task failed."), :error)
-          end
-          add_flash(_("Queued live migration of Instance \"%{name}\"") % {:name => vm.name})
-        else
-          add_flash(_("Unable to live migrate Instance \"%{name}\": %{details}") % {
-            :name    => vm.name,
-            :details => vm.unsupported_reason(:live_migrate)
-          }, :error)
-        end
-      end
-    end
-    @sb[:action] = nil
-    if @sb[:explorer]
-      replace_right_cell
-    else
-      session[:flash_msgs] = @flash_array.dup
-      javascript_redirect previous_breadcrumb_url
-    end
-  end
-
-  def evacuate
-    assert_privileges("instance_evacuate")
-    @record ||= VmOrTemplate.find_by_id(params[:rec_id])
-    drop_breadcrumb(
-      :name => _("Evacuate Instances"),
-      :url  => "/vm_cloud/evacuate"
-    ) unless @explorer
-    @sb[:explorer] = @explorer
-    @in_a_form = true
-    @evacuate = true
-
-    @evacuate_items = VmOrTemplate.find(session[:evacuate_items]).sort_by(&:name)
-    build_targets_hash(@evacuate_items)
-    @view = get_db_view(VmOrTemplate)
-    @view.table = MiqFilter.records2table(@evacuate_items, @view.cols + ['id'])
-
-    render :action => "show" unless @explorer
-  end
-
-  def evacuatevms
-    assert_privileges("instance_evacuate")
-    recs = find_checked_ids_with_rbac(VmOrTemplate)
-    recs = [params[:id].to_i] if recs.blank?
-    session[:evacuate_items] = recs
-    if @explorer
-      evacuate
-      @refresh_partial = "vm_common/evacuate"
-    else
-      javascript_redirect :controller => 'vm', :action => 'evacuate', :escape => false
-    end
-  end
-  alias instance_evacuate evacuatevms
-
-  def evacuate_vm
-    assert_privileges("instance_evacuate")
-    case params[:button]
-    when "cancel"
-      add_flash(_("Evacuation of Instances was cancelled by the user"))
-    when "submit"
-      @evacuate_items = VmOrTemplate.find(session[:evacuate_items]).sort_by(&:name)
-      @evacuate_items.each do |vm|
-        if vm.supports_evacuate?
-          options = {
-            :hostname          => params['auto_select_host']  == 'on' ? nil : params['destination_host'],
-            :on_shared_storage => params['on_shared_storage'] == 'on',
-            :admin_password    => params['on_shared_storage'] == 'on' ? nil : params['admin_password']
-          }
-          task_id = vm.class.evacuate_queue(session[:userid], vm, options)
-          unless task_id.kind_of?(Integer)
-            add_flash(_("Instance evacuation task failed."), :error)
-          end
-          add_flash(_("Queued evacuation of Instance \"%{name}\"") % {:name => vm.name})
-        else
-          add_flash(_("Unable to evacuate Instance \"%{name}\": %{details}") % {
-            :name    => vm.name,
-            :details => vm.unsupported_reason(:evacuate)
-          }, :error)
-        end
-      end
-    end
-    @sb[:action] = nil
-    if @sb[:explorer]
-      replace_right_cell
-    else
-      session[:flash_msgs] = @flash_array.dup
-      javascript_redirect previous_breadcrumb_url
-    end
-  end
-
-  def evacuate_form_fields
-    assert_privileges("instance_evacuate")
-    @record = find_record_with_rbac(VmOrTemplate, params[:id])
-    hosts = []
-    unless @record.ext_management_system.nil?
-      begin
-        connection = @record.ext_management_system.connect
-        current_hostname = connection.handled_list(:servers).find do |s|
-          s.name == @record.name
-        end.os_ext_srv_attr_hypervisor_hostname
-        hosts = connection.hosts.select { |h| h.service_name == "compute" && h.host_name != current_hostname }.map do |h|
-          {:name => h.host_name, :id => h.host_name}
-        end
-      rescue
-        hosts = []
-      end
-    end
-    render :json => {
-      :hosts => hosts
-    }
-  end
-
-  def associate_floating_ip_vms
-    assert_privileges("instance_associate_floating_ip")
-    recs = find_checked_items
-    recs = [params[:id].to_i] if recs.blank?
-    @record = find_record_with_rbac(VmCloud, recs.first)
-    if @record.supports_associate_floating_ip? && @record.ext_management_system.present?
-      if @explorer
-        associate_floating_ip
-        @refresh_partial = "vm_common/associate_floating_ip"
-      else
-        render :update do |page|
-          page << javascript_prologue
-          page.redirect_to :controller => 'vm',
-                           :action     => 'associate_floating_ip',
-                           :rec_id     => @record.id,
-                           :escape     => false
-        end
-      end
-    else
-      add_flash(_("Unable to associate Floating IP with Instance \"%{name}\": %{details}") % {
-        :name    => @record.name,
-        :details => @record.unsupported_reason(:associate_floating_ip)}, :error)
-    end
-  end
-  alias instance_associate_floating_ip associate_floating_ip_vms
-
-  def associate_floating_ip
-    assert_privileges("instance_associate_floating_ip")
-    @record ||= find_record_with_rbac(VmCloud, params[:rec_id])
-    drop_breadcrumb(
-      :name => _("Associate Floating IP with Instance '%{name}'") % {:name => @record.name},
-      :url  => "/vm_cloud/associate_floating_ip"
-    ) unless @explorer
-    @sb[:explorer] = @explorer
-    @in_a_form = true
-    @associate_floating_ip = true
-    render :action => "show" unless @explorer
-  end
-
-  def associate_floating_ip_form_fields
-    assert_privileges("instance_associate_floating_ip")
-    @record = find_record_with_rbac(VmCloud, params[:id])
-    floating_ips = []
-    unless @record.cloud_tenant.nil?
-      floating_ips = @record.cloud_tenant.floating_ips
-    end
-    render :json => {
-      :floating_ips => floating_ips
-    }
-  end
-
-  def associate_floating_ip_vm
-    assert_privileges("instance_associate_floating_ip")
-    @record = find_record_with_rbac(VmCloud, params[:id])
-    case params[:button]
-    when "cancel"
-      add_flash(_("Association of Floating IP with Instance \"%{name}\" was cancelled by the user") % {:name => @record.name})
-      @record = @sb[:action] = nil
-    when "submit"
-      if @record.supports_associate_floating_ip?
-        floating_ip = params[:floating_ip]
-        begin
-          @record.associate_floating_ip_queue(session[:userid], floating_ip)
-          add_flash(_("Associating Floating IP %{address} with Instance \"%{name}\"") % {
-            :address => floating_ip,
-            :name    => @record.name})
-        rescue => ex
-          add_flash(_("Unable to associate Floating IP %{address} with Instance \"%{name}\": %{details}") % {
-            :address => floating_ip,
-            :name    => @record.name,
-            :details => get_error_message_from_fog(ex.to_s)}, :error)
-        end
-      else
-        add_flash(_("Unable to associate Floating IP with Instance \"%{name}\": %{details}") % {
-          :name    => @record.name,
-          :details => @record.unsupported_reason(:associate_floating_ip)}, :error)
-      end
-      params[:id] = @record.id.to_s # reset id in params for show
-      @record = nil
-      @sb[:action] = nil
-    end
-    if @sb[:explorer]
-      replace_right_cell
-    else
-      session[:flash_msgs] = @flash_array.dup
-      render :update do |page|
-        page << javascript_prologue
-        page.redirect_to previous_breadcrumb_url
-      end
-    end
-  end
-
-  def disassociate_floating_ip_vms
-    assert_privileges("instance_disassociate_floating_ip")
-    recs = find_checked_items
-    recs = [params[:id].to_i] if recs.blank?
-    @record = find_record_with_rbac(VmCloud, recs.first)
-    if @record.supports_disassociate_floating_ip? && @record.ext_management_system.present?
-      if @explorer
-        disassociate_floating_ip
-        @refresh_partial = "vm_common/disassociate_floating_ip"
-      else
-        render :update do |page|
-          page << javascript_prologue
-          page.redirect_to :controller => 'vm',
-                           :action     => 'disassociate_floating_ip',
-                           :rec_id     => @record.id,
-                           :escape     => false
-        end
-      end
-    else
-      add_flash(_("Unable to disassociate Floating IP from Instance \"%{name}\": %{details}") % {
-        :name    => @record.name,
-        :details => @record.unsupported_reason(:disassociate_floating_ip)}, :error)
-    end
-  end
-  alias instance_disassociate_floating_ip disassociate_floating_ip_vms
-
-  def disassociate_floating_ip
-    assert_privileges("instance_disassociate_floating_ip")
-    @record ||= VmCloud.find_by_id(params[:rec_id])
-    drop_breadcrumb(
-      :name => _("Disssociate Floating IP from Instance '%{name}'") % {:name => @record.name},
-      :url  => "/vm_cloud/disassociate_floating_ip"
-    ) unless @explorer
-    @sb[:explorer] = @explorer
-    @in_a_form = true
-    @live_migrate = true
-    render :action => "show" unless @explorer
-  end
-
-  def disassociate_floating_ip_form_fields
-    assert_privileges("instance_disassociate_floating_ip")
-    @record = find_record_with_rbac(VmCloud, params[:id])
-    floating_ips = []
-    unless @record.ext_management_system.nil?
-      @record.floating_ips.each do |floating_ip|
-        floating_ips << floating_ip
-      end
-    end
-    render :json => {
-      :floating_ips => floating_ips
-    }
-  end
-
-  def disassociate_floating_ip_vm
-    assert_privileges("instance_disassociate_floating_ip")
-    @record = find_record_with_rbac(VmCloud, params[:id])
-    case params[:button]
-    when "cancel"
-      add_flash(_("Disassociation of Floating IP from Instance \"%{name}\" was cancelled by the user") % {:name => @record.name})
-      @record = @sb[:action] = nil
-    when "submit"
-      if @record.supports_disassociate_floating_ip?
-        floating_ip = params[:floating_ip]
-        begin
-          @record.disassociate_floating_ip_queue(session[:userid], floating_ip)
-          add_flash(_("Disassociating Floating IP %{address} from Instance \"%{name}\"") % {
-            :address => floating_ip,
-            :name    => @record.name})
-        rescue => ex
-          add_flash(_("Unable to disassociate Floating IP %{address} from Instance \"%{name}\": %{details}") % {
-            :address => floating_ip,
-            :name    => @record.name,
-            :details => get_error_message_from_fog(ex.to_s)}, :error)
-        end
-      else
-        add_flash(_("Unable to disassociate Floating IP from Instance \"%{name}\": %{details}") % {
-          :name    => @record.name,
-          :details => @record.unsupported_reason(:disassociate_floating_ip)}, :error)
-      end
-      params[:id] = @record.id.to_s # reset id in params for show
-      @record = nil
-      @sb[:action] = nil
-    end
-    if @sb[:explorer]
-      replace_right_cell
-    else
-      session[:flash_msgs] = @flash_array.dup
-      render :update do |page|
-        page << javascript_prologue
-        page.redirect_to previous_breadcrumb_url
-      end
-    end
-  end
-
-  def vm_right_size
-    assert_privileges(params[:pressed])
-    # check to see if coming from show_list or drilled into vms from another CI
-    rec_cls = "vm"
-    recs = params[:display] ? find_checked_ids_with_rbac(VmOrTemplate) : [find_id_with_rbac(VmOrTemplate, params[:id]).to_i]
-    if recs.length < 1
-      add_flash(_("One or more %{model} must be selected to Right-Size Recommendations") %
-        {:model => ui_lookup(:table => request.parameters[:controller])}, :error)
-      @refresh_div = "flash_msg_div"
-      @refresh_partial = "layouts/flash_msg"
-      return
-    else
-      if VmOrTemplate.includes_template?(recs)
-        add_flash(_("Right-Size Recommendations does not apply to selected %{model}") %
-          {:model => ui_lookup(:table => "miq_template")}, :error)
-        javascript_flash(:scroll_top => true)
-        return
-      end
-    end
-    if @explorer
-      @refresh_partial = "vm_common/right_size"
-      right_size
-      replace_right_cell if @orig_action == "x_history"
-    else
-      if role_allows?(:feature => "vm_right_size")
-        javascript_redirect :controller => rec_cls.to_s, :action => 'right_size', :id => recs[0], :escape => false # redirect to build the ownership screen
-      else
-        head :ok
-      end
-    end
-  end
-  alias_method :instance_right_size, :vm_right_size
-
-  # Assign/unassign ownership to a set of objects
-  def reconfigure
-    @sb[:explorer] = true if @explorer
-    @request_id = nil
-    @in_a_form = @reconfigure = true
-    drop_breadcrumb(:name => _("Reconfigure"), :url => "/vm_common/reconfigure")
-    if params[:rec_ids]
-      @reconfigure_items = params[:rec_ids]
-    end
-    if params[:req_id]
-      @request_id = params[:req_id]
-    end
-    @reconfigitems = Vm.find(@reconfigure_items).sort_by(&:name) # Get the db records
-    build_targets_hash(@reconfigitems)
-    @force_no_grid_xml   = true
-    @view, @pages = get_view(Vm, :view_suffix => "VmReconfigureRequest", :where_clause => ["vms.id IN (?)", @reconfigure_items])  # Get the records (into a view) and the paginator
-    get_reconfig_limits
-    unless @explorer
-      render :action => "show"
-    end
-  end
-
-  def reconfigure_update
-    case params[:button]
-    when "cancel"
-      add_flash(_("VM Reconfigure Request was cancelled by the user"))
-      if @sb[:explorer]
-        @sb[:action] = nil
-        replace_right_cell
-      else
-        session[:flash_msgs] = @flash_array
-        javascript_redirect previous_breadcrumb_url
-      end
-    when "submit"
-      options = {:src_ids => params[:objectIds]}
-      if params[:cb_memory] == 'true'
-        options[:vm_memory] = params[:memory_type] == "MB" ? params[:memory] : (params[:memory].to_i.zero? ? params[:memory] : params[:memory].to_i * 1024)
-      end
-      if params[:cb_cpu] == 'true'
-        options[:cores_per_socket]  = params[:cores_per_socket_count].nil? ? 1 : params[:cores_per_socket_count].to_i
-        options[:number_of_sockets] = params[:socket_count].nil? ? 1 : params[:socket_count].to_i
-        vccores = params[:cores_per_socket_count] == 0 ? 1 : params[:cores_per_socket_count]
-        vsockets = params[:socket_count] == 0 ? 1 : params[:socket_count]
-        options[:number_of_cpus] = vccores.to_i * vsockets.to_i
-      end
-      # set the disk_add and disk_remove options
-      if params[:vmAddDisks]
-        params[:vmAddDisks].values.each do |p|
-          p.transform_values!{ |v|
-            case v
-              when 'true'
-                true
-              when 'false'
-                false
-              else
-                v
-            end }
-        end
-        options[:disk_add] = params[:vmAddDisks].values
-      end
-
-      if params[:vmRemoveDisks]
-        params[:vmRemoveDisks].values.each do |p|
-          p.transform_values!{ |v|
-            case v
-            when 'true'
-              true
-            when 'false'
-              false
-            else
-              v
-            end }
-        end
-        options[:disk_remove] = params[:vmRemoveDisks].values
-      end
-
-      if(params[:id] && params[:id] != 'new')
-        @request_id = params[:id]
-      end
-
-      VmReconfigureRequest.make_request(@request_id, options, current_user)
-      flash = _("VM Reconfigure Request was saved")
-
-      if role_allows?(:feature => "miq_request_show_list", :any => true)
-        javascript_redirect :controller => 'miq_request', :action => 'show_list', :flash_msg => flash
-      else
-        url = previous_breadcrumb_url.split('/')
-        javascript_redirect :controller => url[1], :action => url[2], :flash_msg => flash
-      end
-
-      if @flash_array
-        javascript_flash
-        return
-      end
-    end
-  end
-
-  def reconfigure_form_fields
-    request_data = ''
-    @request_id, request_data = params[:id].split(/\s*,\s*/, 2)
-    @reconfigure_items = request_data.split(/\s*,\s*/)
-    request_hash = build_reconfigure_hash
-    render :json => request_hash
   end
 
   # edit single selected Object
@@ -1016,7 +61,7 @@ module ApplicationController::CiProcessing
     end
   end
 
-  # copy single selected Object
+  # copy single selected Object # FIXME: dead code?
   def copy_record
     obj = find_checked_items
     @refresh_partial = "copy"
@@ -1454,6 +499,14 @@ module ApplicationController::CiProcessing
     end
   end
 
+  def get_error_message_from_fog(exception)
+    exception_string = exception.to_s
+    matched_message = exception_string.match(/message\\\": \\\"(.*)\\\", /)
+    matched_message ? matched_message[1] : exception_string
+  end
+
+  private
+
   def process_elements(elements, klass, task, display_name = nil, order_field = nil)
     order_field ||= %w(name description title).find do |field|
                       klass.column_names.include?(field)
@@ -1492,14 +545,6 @@ module ApplicationController::CiProcessing
       end
     end
   end
-
-  def get_error_message_from_fog(exception)
-    exception_string = exception.to_s
-    matched_message = exception_string.match(/message\\\": \\\"(.*)\\\", /)
-    matched_message ? matched_message[1] : exception_string
-  end
-
-  private ############################
 
   def explorer_controller?
     %w(vm_cloud vm_infra vm_or_template infra_networking).include?(controller_name)
@@ -1581,192 +626,6 @@ module ApplicationController::CiProcessing
 
   def breadcrumb_name(_model)
     ui_lookup(:models => self.class.model.name)
-  end
-
-  # Reconfigure selected VMs
-  def reconfigurevms
-    assert_privileges(params[:pressed])
-    @request_id = nil
-    # check to see if coming from show_list or drilled into vms from another CI
-    rec_cls = "vm"
-    recs = []
-    # if coming in to edit from miq_request list view
-    if !session[:checked_items].nil? && (@lastaction == "set_checked_items" || params[:pressed] == "miq_request_edit")
-      @request_id = params[:id]
-      recs = session[:checked_items]
-    elsif !params[:id] || params[:pressed] == 'vm_reconfigure'
-      recs = find_checked_ids_with_rbac(VmOrTemplate)
-    end
-    if recs.blank?
-      recs = [find_id_with_rbac(VmOrTemplate, params[:id]).to_i]
-    end
-    if recs.length < 1
-      add_flash(_("One or more %{model} must be selected to Reconfigure") %
-        {:model => Dictionary.gettext(db.to_s, :type => :model, :notfound => :titleize, :plural => true)}, :error)
-      javascript_flash(:scroll_top => true)
-      return
-    else
-
-      if VmOrTemplate.includes_template?(recs)
-        add_flash(_("Reconfigure does not apply because you selected at least one %{model}") %
-          {:model => ui_lookup(:table => "miq_template")}, :error)
-        javascript_flash(:scroll_top => true)
-        return
-      end
-      unless VmOrTemplate.reconfigurable?(recs)
-        add_flash(_("Reconfigure does not apply because you selected at least one un-reconfigurable VM"), :error)
-        javascript_flash(:scroll_top => true)
-        return
-      end
-      @reconfigure_items = recs.collect(&:to_i)
-    end
-    if @explorer
-      reconfigure
-      session[:changed] = true  # need to enable submit button when screen loads
-      @refresh_partial = "vm_common/reconfigure"
-    else
-      if role_allows?(:feature => "vm_reconfigure")
-        javascript_redirect :controller => rec_cls.to_s, :action => 'reconfigure', :req_id => @request_id, :rec_ids => @reconfigure_items, :escape => false # redirect to build the ownership screen
-      else
-        head :ok
-      end
-    end
-  end
-  alias_method :image_reconfigure, :reconfigurevms
-  alias_method :instance_reconfigure, :reconfigurevms
-  alias_method :vm_reconfigure, :reconfigurevms
-  alias_method :miq_template_reconfigure, :reconfigurevms
-
-  def get_reconfig_info
-    @reconfigureitems = Vm.find(@reconfigure_items).sort_by(&:name)
-    # set memory to nil if multiple items were selected with different mem_cpu values
-    memory = @reconfigureitems.first.mem_cpu
-    memory = nil unless @reconfigureitems.all? { |vm| vm.mem_cpu == memory }
-
-    socket_count = @reconfigureitems.first.num_cpu
-    socket_count = '' unless @reconfigureitems.all? { |vm| vm.num_cpu == socket_count }
-
-    cores_per_socket = @reconfigureitems.first.cpu_cores_per_socket
-    cores_per_socket = '' unless @reconfigureitems.all? { |vm| vm.cpu_cores_per_socket == cores_per_socket }
-    memory, memory_type = reconfigure_calculations(memory)
-
-    # if only one vm that supports disk reconfiguration is selected, get the disks information
-    vmdisks = []
-    @reconfigureitems.first.hardware.disks.each do |disk|
-      next if disk.device_type != 'disk'
-      dsize, dunit = reconfigure_calculations(disk.size / (1024 * 1024))
-      vmdisks << {:hdFilename  => disk.filename,
-                  :hdType      => disk.disk_type,
-                  :hdMode      => disk.mode,
-                  :hdSize      => dsize,
-                  :hdUnit      => dunit,
-                  :add_remove  => '',
-                  :cb_bootable => disk.bootable}
-    end
-
-    {:objectIds              => @reconfigure_items,
-     :memory                 => memory,
-     :memory_type            => memory_type,
-     :socket_count           => socket_count.to_s,
-     :cores_per_socket_count => cores_per_socket.to_s,
-     :disks                  => vmdisks}
-  end
-
-  def supports_reconfigure_disks?
-    @reconfigitems && @reconfigitems.size == 1 && @reconfigitems.first.supports_reconfigure_disks?
-  end
-
-  def get_reconfig_limits
-    @reconfig_limits = VmReconfigureRequest.request_limits(:src_ids => @reconfigure_items)
-    mem1, fmt1 = reconfigure_calculations(@reconfig_limits[:min__vm_memory])
-    mem2, fmt2 = reconfigure_calculations(@reconfig_limits[:max__vm_memory])
-    @reconfig_memory_note = "Between #{mem1}#{fmt1} and #{mem2}#{fmt2}"
-
-    @socket_options = []
-    @reconfig_limits[:max__number_of_sockets].times do |tidx|
-      idx = tidx + @reconfig_limits[:min__number_of_sockets]
-      @socket_options.push(idx) if idx <= @reconfig_limits[:max__number_of_sockets]
-    end
-
-    @cores_options = []
-    @reconfig_limits[:max__cores_per_socket].times do |tidx|
-      idx = tidx + @reconfig_limits[:min__cores_per_socket]
-      @cores_options.push(idx) if idx <= @reconfig_limits[:max__cores_per_socket]
-    end
-  end
-
-  # Build the reconfigure data hash
-  def build_reconfigure_hash
-    @req = nil
-    @reconfig_values = {}
-    if @request_id == 'new'
-      @reconfig_values = get_reconfig_info
-    else
-      @req = MiqRequest.find_by_id(@request_id)
-      @reconfig_values[:src_ids] = @req.options[:src_ids]
-      @reconfig_values[:memory], @reconfig_values[:memory_type] = @req.options[:vm_memory] ? reconfigure_calculations(@req.options[:vm_memory]) : ['','']
-      @reconfig_values[:cores_per_socket_count] = @req.options[:cores_per_socket] ? @req.options[:cores_per_socket].to_s : ''
-      @reconfig_values[:socket_count] = @req.options[:number_of_sockets] ? @req.options[:number_of_sockets].to_s : ''
-      # check if there is only one VM that supports disk reconfiguration
-
-      @reconfig_values[:disk_add] = @req.options[:disk_add]
-      @reconfig_values[:disk_remove] = @req.options[:disk_remove]
-      vmdisks = []
-      if @req.options[:disk_add]
-        @req.options[:disk_add].each do |disk|
-          adsize, adunit = reconfigure_calculations(disk[:disk_size_in_mb])
-          vmdisks << {:hdFilename   => disk[:disk_name],
-                      :hdType       => disk[:thin_provisioned] ? 'thin' : 'thick',
-                      :hdMode       => disk[:persistent] ? 'persistent' : 'nonpersistent',
-                      :hdSize       => adsize.to_s,
-                      :hdUnit       => adunit,
-                      :cb_dependent => disk[:dependent],
-                      :cb_bootable  => disk[:bootable],
-                      :add_remove   => 'add'}
-        end
-      end
-
-      reconfig_item = Vm.find(@reconfigure_items)
-      if reconfig_item
-        reconfig_item.first.hardware.disks.each do |disk|
-          next if disk.device_type != 'disk'
-          removing = ''
-          delbacking = false
-          if disk.filename && @req.options[:disk_remove]
-            @req.options[:disk_remove].each do |remdisk|
-              if remdisk[:disk_name] == disk.filename
-                removing = 'remove'
-                delbacking = remdisk[:delete_backing]
-              end
-            end
-          end
-          dsize, dunit = reconfigure_calculations(disk.size / (1024 * 1024))
-          vmdisks << {:hdFilename     => disk.filename,
-                      :hdType         => disk.disk_type.to_s,
-                      :hdMode         => disk.mode.to_s,
-                      :hdSize         => dsize.to_s,
-                      :hdUnit         => dunit.to_s,
-                      :delete_backing => delbacking,
-                      :cb_bootable    => disk.bootable,
-                      :add_remove     => removing}
-        end
-      end
-      @reconfig_values[:disks] = vmdisks
-    end
-
-    @reconfig_values[:cb_memory] = !!(@req && @req.options[:vm_memory])       # default for checkbox is false for new request
-    @reconfig_values[:cb_cpu] =  !!(@req && ( @req.options[:number_of_sockets] || @req.options[:cores_per_socket]))     # default for checkbox is false for new request
-    @reconfig_values
-  end
-
-  def reconfigure_calculations(mbsize)
-    humansize = mbsize
-    fmt = "MB"
-    if mbsize.to_i > 1024 && mbsize.to_i % 1024 == 0
-      humansize = mbsize.to_i / 1024
-      fmt = "GB"
-    end
-    return humansize.to_s, fmt
   end
 
   # Common item button handler routines
